Add session mixin consolidation
